% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{alloc_cache}
\alias{alloc_cache}
\title{Allocate/deallocate Cache}
\usage{
alloc_cache(cache, force = F)
}
\description{
This system allows reading from the cache on the filesystem
while storing the data in memory when possible. This reduces
filesystem traffic while minimizing risk that cache will be
changed while code is executing.
}
\details{
You shouldn't have to use these functions unless you are doing
a lot of cache work. They are called behind the scenes in the
`target` function, and any other function that accesses the cache.
}
