% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apple.R
\docType{data}
\name{apple}
\alias{apple}
\title{Quarterly reporting data by Apple Inc (2015-2018)}
\format{Data frame with columns
\describe{
\item{Year}{Year}
\item{Quarter}{Quarter}
\item{Product}{Product line}
\item{Units}{Units sold in quarter (millions)}
\item{Revenue}{Revenue (millions of USD)}
}}
\source{
<>
}
\usage{
apple
}
\description{
Quarterly reporting data by Apple Inc (2015-2018)
}
\keyword{datasets}
