% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/areas.R
\docType{data}
\name{areas}
\alias{areas}
\title{Areas of some large countries}
\format{Data frame with columns
\describe{
\item{country}{Name of country}
\item{area}{Land area}
}}
\source{
<https://www.worldometers.info/geography/largest-countries-in-the-world/>
}
\usage{
areas
}
\description{
Areas of some large countries
}
\keyword{datasets}
