% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cars2018.R
\docType{data}
\name{cars2018}
\alias{cars2018}
\title{Information about 2018 model year vehicles}
\format{Data frame with columns
\describe{
\item{Model}{Model Name}
\item{ID}{ID}
\item{Displacement}{Engine displacement}
\item{Cylinders}{Number of cylinders}
\item{Gears}{Number of Gears}
\item{Transmission}{Transmission: Automatic, CVT, or Manual}
\item{MPG}{Miles per gallon}
\item{Aspiration}{Naturally Aspirated or Turbocharged/Supercharged }
\item{LTC}{Lockup Torque Converter: Y or N}
\item{Drive}{Drive system}
\item{MaxEthanol}{Maximum ethanol in ethanol fuel mixture}
\item{Fuel}{Fuel Type}
\item{Intake}{Intake valves per cylinder}
\item{Exhaust}{Exhaust valves per cylinder}
\item{FuelInjection}{Direct ignition or Multipoint/sequential ignition }
}}
\source{
<https://supervised-ml-course.netlify.com/>
}
\usage{
cars2018
}
\description{
Information about 2018 model year vehicles
}
\keyword{datasets}
