% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi_emps.R
\docType{data}
\name{chi_emps}
\alias{chi_emps}
\title{Employees of the City of Chicago}
\format{Data frame with columns
\describe{
\item{Names}{Employee names}
\item{Titles}{Job titles}
\item{Dept}{Department of the city government}
\item{FullPart}{Employee's status: full-time or part-time}
\item{SalHour}{Whether the employee is salaried or hourly-paid}
\item{TypicalHours}{Typical hours per week for those who are hourly-paid}
\item{AnnSalary}{Annual Salary for those who are salaried}
\item{HourlyRate}{Hourly rate for those who are hourly-paid}
}}
\source{
<https://data.cityofchicago.org/Administration-Finance/Current-Employee-Names-Salaries-and-Position-Title/xzkq-xp2w>
}
\usage{
chi_emps
}
\description{
Employees of the City of Chicago
}
\keyword{datasets}
