% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groceries.R
\docType{data}
\name{groceries}
\alias{groceries}
\title{Prices of selected grocery items in Trinidad and Tobago in November 2018}
\format{Data frame with columns
\describe{
\item{No.}{Item identifier as assigned by Ministry of Trade and Information}
\item{ITEMS}{Name of item}
\item{BRAND}{Brand of item}
\item{SIZE}{Item size}
\item{Outlet}{Name of the grocery}
\item{Town}{Location of the grocery}
\item{Region}{Location: Central, East, North, South, Tobago, West}
\item{Price}{Item price}
}}
\source{
<https://tradeind.gov.tt/tag/supermarket-pricelist/>
}
\usage{
groceries
}
\description{
Prices of selected grocery items in Trinidad and Tobago in November 2018
}
\keyword{datasets}
