% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highrises.R
\docType{data}
\name{inspections}
\alias{inspections}
\title{Results of inspections of highrise buildings in Toronto (2017 - 2020)}
\format{Data frame with columns
\describe{
\item{id}{Key column}
\item{Opened}{Date the issue was opened}
\item{Closed}{Date the issue was closed}
\item{Address}{Street Address}
\item{Code}{Violation code}
\item{Desc}{Description of violation}
\item{Lat}{Latitude of the building's location}
\item{Lon}{Latitude of the building's location}
\item{Ward}{Electoral ward of Toronto/Ontario}
}}
\source{
<https://open.toronto.ca/dataset/highrise-residential-fire-inspection-results/>
}
\usage{
inspections
}
\description{
Results of inspections of highrise buildings in Toronto (2017 - 2020)
}
\keyword{datasets}
