% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inventories.R
\docType{data}
\name{inventories}
\alias{inventories}
\title{US Manufacturing and Trade inventories 1992 - 2019}
\format{Data frame with columns
\describe{
\item{Month}{Calendar month}
\item{Year}{Year}
\item{Total}{Total Inventory (000s of USD)}
\item{Manufacturing}{Inventory in the Manufacturing sector (000s of USD)}
\item{Retail}{Inventory in the retail sector (000s of USD)}
\item{Wholesale}{Inventory in the wholesale sector (000s of USD)}
}}
\source{
<https://www.census.gov/mtis/www/data/text/timeseries2.xlsx>
}
\usage{
inventories
}
\description{
US Manufacturing and Trade inventories 1992 - 2019
}
\keyword{datasets}
