% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movies.R
\docType{data}
\name{movies}
\alias{movies}
\title{Budget and box office numbers for movies}
\format{Data frame with columns
\describe{
\item{release_date}{Date the movie was released}
\item{movie}{Name of movie}
\item{production_budget}{Movie budget (USD)}
\item{domestic_gross}{Revenue of movie in the USA (USD)}
\item{worldwide_gross}{Revenue of movie globally (USD)}
\item{distributor}{Studio distributing the movie}
\item{mpaa_rating}{Movie rating}
\item{genre}{Movie genre}
}}
\source{
<https://github.com/rfordatascience/tidytuesday/blob/master/data/2018/2018-10-23/movie_profit.csv originally from the-numbers.com>
}
\usage{
movies
}
\description{
Budget and box office numbers for movies
}
\keyword{datasets}
