% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba1819.R
\docType{data}
\name{nba1819}
\alias{nba1819}
\title{Games of the 2018-2019 NBA Season}
\format{An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 1312 rows and 9 columns.}
\source{
<https://www.basketball-reference.com/leagues/NBA_2020_games.html>
}
\usage{
nba1819
}
\description{
Games of the 2018-2019 NBA Season
}
\keyword{datasets}
