% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nyc311.R
\docType{data}
\name{nyc311}
\alias{nyc311}
\title{NYC 311 Service Requests for Jan 1-14 2020}
\format{Data frame with columns
\describe{
\item{Key}{Case identifier}
\item{Created}{Date Created}
\item{Closed}{Date Closed}
\item{Agency}{NYC Agency Acronym}
\item{Type}{Type of complaint}
\item{LocType}{Type of location}
\item{Zip}{ZIP code of complaint}
\item{Status}{Complaint status: Pending, Assigned, Started, In Progress, Open or Closed }
\item{Borough}{Borough of NYC}
\item{Latitude}{Latitude}
\item{Longitude}{Longitude}
}}
\source{
<https://nycopendata.socrata.com/Social-Services/311-Service-Requests-from-2010-to-Present/erm2-nwe9>
}
\usage{
nyc311
}
\description{
NYC 311 Service Requests for Jan 1-14 2020
}
\keyword{datasets}
