% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nyc_sat10.R
\docType{data}
\name{nyc_sat10}
\alias{nyc_sat10}
\title{Avg SAT scores at NYC schools in 2010}
\format{Data frame with columns
\describe{
\item{DBN}{School Identifier}
\item{SCHOOL}{School name}
\item{STUDENTS}{Number of students}
\item{READING}{Average score on the SAT Reading test}
\item{MATH}{Average score on the SAT Math test}
\item{WRITING}{Average score on the SAT Writing test}
}}
\source{
<https://data.cityofnewyork.us/Education/SAT-College-Board-2010-School-Level-Results/zt9s-n5aj>
}
\usage{
nyc_sat10
}
\description{
Avg SAT scores at NYC schools in 2010
}
\keyword{datasets}
