% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nyc_sat12.R
\docType{data}
\name{nyc_sat12}
\alias{nyc_sat12}
\title{Avg SAT scores at NYC schools in 2012}
\format{Data frame with columns
\describe{
\item{DBN}{School Identifier}
\item{SCHOOL}{School name}
\item{STUDENTS}{Number of students}
\item{READING}{Average score on the SAT Reading test}
\item{MATH}{Average score on the SAT Math test}
\item{WRITING}{Average score on the SAT Writing test}
}}
\source{
<https://data.cityofnewyork.us/Education/2012-SAT-Results/f9bf-2cp4>
}
\usage{
nyc_sat12
}
\description{
Avg SAT scores at NYC schools in 2012
}
\keyword{datasets}
