% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populations.R
\docType{data}
\name{populations}
\alias{populations}
\title{Populations of some large countries}
\format{Data frame with columns
\describe{
\item{country}{Name of country}
\item{pop}{Population (millions)}
}}
\source{
</>
}
\usage{
populations
}
\description{
Populations of some large countries
}
\keyword{datasets}
