% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purify.R
\name{purify_function}
\alias{purify_function}
\title{PURITY}
\usage{
purify_function(func, ignore_arg_defaults = T)
}
\description{
What is allowed inside of target methods:
- Explicitly specified dependencies
  - Objects                                Invalidated through inequality
  - File paths                             Invalidated through file modified dates
  - Other targets                          Invalidated recursively
  - Dirty functions from devtools shim /   Must specify cache invalidation method
    method source environment
- Functions from packages                  Invalidated through package versioning
- Pure functions from devtools shim /      Invalidated through recursive code analysis
  method source environment
}
\details{
What is not allowed inside of target methods:
- Unspecified non-function objects from method source environment
- Loading from unspecified file paths
- Other targets
}
