% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rafa_novak.R
\docType{data}
\name{rafa_novak}
\alias{rafa_novak}
\title{Rafael Nadal vs Novak Djokovic head to head}
\format{Data frame with columns
\describe{
\item{Year}{Year}
\item{Event}{Event}
\item{Location}{Geographical location}
\item{Surface}{Type of surface (court)}
\item{RND}{Tournament Round}
\item{Winner}{Name of winner}
\item{Result}{Text field with match scores}
\item{Loser}{Name of loser}
}}
\source{
<https://www.atptour.com/en/players/fedex-head-2-head/novak-djokovic-vs-rafael-nadal/D643/N409>
}
\usage{
rafa_novak
}
\description{
Rafael Nadal vs Novak Djokovic head to head
}
\keyword{datasets}
