% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-words.R
\name{read_words_from_sheet}
\alias{read_words_from_sheet}
\title{Read word sheet}
\usage{
read_words_from_sheet(topic = NULL)
}
\arguments{
\item{topic}{Character string naming one of the topics
available from the sheet. Omit argument to get a list of those topics.}
}
\description{
Read word sheet
}
