% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sales.R
\docType{data}
\name{sales}
\alias{sales}
\title{US Manufacturing and Trade Sales 1992 - 2019}
\format{Data frame with columns
\describe{
\item{Month}{Calendar month}
\item{Year}{Year}
\item{Total}{Total Sales (000s of USD)}
\item{Manufacturing}{Sales in the Manufacturing sector (000s of USD)}
\item{Retail}{Sales in the retail sector (000s of USD)}
\item{Wholesale}{Sales in the wholesale sector (000s of USD)}
}}
\source{
<https://www.census.gov/mtis/www/data/text/timeseries2.xlsx>
}
\usage{
sales
}
\description{
US Manufacturing and Trade Sales 1992 - 2019
}
\keyword{datasets}
