% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/us_unemp.R
\docType{data}
\name{us_unemp}
\alias{us_unemp}
\title{US unemployment rates 1948 - 2019}
\format{Data frame with columns
\describe{
\item{Year}{Year}
\item{Jan}{Unemployment rate during January of the specified year}
\item{Feb}{Unemployment rate during February of the specified year}
\item{Mar}{Unemployment rate during March of the specified year}
\item{Apr}{Unemployment rate during April of the specified year}
\item{May}{Unemployment rate during May of the specified year}
\item{Jun}{Unemployment rate during June of the specified year}
\item{Jul}{Unemployment rate during July of the specified year}
\item{Aug}{Unemployment rate during August of the specified year}
\item{Sep}{Unemployment rate during September of the specified year}
\item{Oct}{Unemployment rate during October of the specified year}
\item{Nov}{Unemployment rate during November of the specified year}
\item{Dec}{Unemployment rate during December of the specified year}
}}
\source{
<https://data.bls.gov/timeseries/LNS14000000>
}
\usage{
us_unemp
}
\description{
US unemployment rates 1948 - 2019
}
\keyword{datasets}
