% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workshop.R
\name{write_answers}
\alias{write_answers}
\title{Write hands on solutions to file.}
\usage{
write_answers(x, output)
}
\arguments{
\item{x}{parsed hands on object}

\item{output}{output file}
}
\description{
Write hands on solutions to file.
}

