% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workshop.R
\name{write_questions}
\alias{write_questions}
\title{Write hands on problems to file.}
\usage{
write_questions(x, output)
}
\arguments{
\item{x}{parsed hands on object}

\item{output}{output file}
}
\description{
Write hands on problems to file.
}

