% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groceries.R
\docType{data}
\name{groceries}
\alias{groceries}
\title{Prices of selected grocery items in Trinidad and Tobago in November 2018}
\format{An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 846 rows and 8 columns.}
\source{
<https://tradeind.gov.tt/tag/supermarket-pricelist/>
}
\usage{
groceries
}
\description{
Prices of selected grocery items in Trinidad and Tobago in November 2018
}
\keyword{datasets}
